+++ public/index.html
@@ -88,7 +88,7 @@
                 <div style="width: 100%; text-align: center">
                     <img
                         src="%PUBLIC_URL%/logo512.png"
-                        alt="obrázok"
+                        alt=""
                         style="aspect-ratio: 1; width: 50%; max-width: 512px; border-radius: 32px; padding: 0; margin: 0"
                     />
                     <h1
