+++ src/App.tsx
@@ -1,482 +1,14 @@
-import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
-import { Navigate, Route, Routes, Link, Outlet } from 'react-router-dom';
-import { Button, Container, Nav, Navbar, Offcanvas } from 'react-bootstrap';
+import React, { useContext } from 'react';
+import { RouterProvider } from 'react-router-dom';
 import './App.scss';
 import { AuthContext } from './store/auth-context';
-import SignInPage from './pages/signIn/SignInPage';
-import UsersPage from './pages/users/UsersPage';
-import UnitCategoriesPage from './pages/unitCategories/UnitCategoriesPage';
-import TagsPage from './pages/tags/TagsPage';
-import ProfilePage from './pages/profile/ProfilePage';
-import CategoryPage from './pages/categories/CategoryPage';
-import RecipeEditPage from './pages/recipes/RecipeEditPage';
-import { authApi, categoryApi } from './utils/apiWrapper';
-import { Api } from './openapi';
-import UserPage from './pages/users/UserPage';
-import CategoriesPage from './pages/categories/CategoriesPage';
-import TagPage from './pages/tags/TagPage';
-import UnitCategoryPage from './pages/unitCategories/UnitCategoryPage';
-import UnitPage from './pages/unitCategories/UnitPage';
-import RecipeSearchPage from './pages/recipes/RecipeSearchPage';
-import RecipeViewPage from './pages/recipes/RecipeViewPage';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-    faArrowRightFromBracket,
-    faList,
-    faPizzaSlice,
-    faScaleBalanced,
-    faTags,
-    faUser,
-    faUsers,
-    faUtensils
-} from '@fortawesome/free-solid-svg-icons';
-import CookieConsent from 'react-cookie-consent';
-import { formatErrorMessage } from './utils/errorMessages';
-import Spinner from './components/UI/Spinner';
-import Modal from './components/UI/Modal';
-import ConfirmationPage from './pages/signIn/ConfirmationPage';
-import ResetPasswordPage from './pages/signIn/ResetPasswordPage';
-import ResetPasswordRequestPage from './pages/signIn/ResetPasswordRequestPage';
-import { useQuery } from '@tanstack/react-query';
-import useRole from './hooks/useRole';
-import { getDefaultSearchParams } from './hooks/useCriteria';
-import { BookmarkContextProvider } from './store/bookmark-context';
-import ConsentPage from './pages/consent/ConsentPage';
-
-const DEFAULT_ICON_STYLE: React.CSSProperties = { width: 20, paddingRight: 8 } as const;
-
-const Navigation: React.FC<{ isLoggedIn: boolean; username: string }> = ({ isLoggedIn, username }) => {
-    const { logout: logoutHandler } = useContext(AuthContext);
-    const [expanded, setExpanded] = useState(false);
-    const [error, setError] = useState<string>();
-    const { hasSome } = useRole();
-
-    const openOffcanvas = useCallback(() => {
-        setExpanded(true);
-    }, []);
-
-    const closeOffcanvas = useCallback(() => {
-        setExpanded(false);
-    }, []);
-
-    const {
-        data: listOfCategories,
-        isFetching: isFetchingListOfCategories,
-        error: listOfCategoriesError
-    } = useQuery({
-        queryKey: ['categories'],
-        queryFn: ({ signal }) =>
-            categoryApi.getCategories({ signal }).then((data) =>
-                data.sort((a, b) =>
-                    a.name.localeCompare(b.name, undefined, {
-                        sensitivity: 'base'
-                    })
-                )
-            ),
-        enabled: isLoggedIn
-    });
-
-    useEffect(() => {
-        if (listOfCategoriesError) {
-            formatErrorMessage(listOfCategoriesError).then((message) => setError(message));
-        }
-    }, [listOfCategoriesError]);
-
-    if (!isLoggedIn) {
-        return null;
-    }
-
-    return (
-        <>
-            <Navbar
-                bg='primary'
-                variant='dark'
-                expand={false}
-                expanded={expanded}
-            >
-                <Container fluid>
-                    <Navbar.Toggle onClick={openOffcanvas} />
-                    <Navbar.Brand
-                        as={Link}
-                        to='/recipes'
-                    >
-                        Kuchárska kniha
-                    </Navbar.Brand>
-                    <Button
-                        onClick={logoutHandler}
-                        aria-label='Odhlásiť sa'
-                    >
-                        <FontAwesomeIcon
-                            className='d-lg-none'
-                            icon={faArrowRightFromBracket}
-                            title='Odhlásiť sa'
-                            aria-hidden='true'
-                        />
-                        <span className='d-none d-lg-inline'>Odhlásiť sa</span>
-                    </Button>
-                    <Navbar.Offcanvas
-                        placement='start'
-                        onHide={closeOffcanvas}
-                    >
-                        <Offcanvas.Header closeButton>
-                            <Offcanvas.Title>{username}</Offcanvas.Title>
-                        </Offcanvas.Header>
-                        <Offcanvas.Body>
-                            <Nav>
-                                <Nav.Link
-                                    to='/profile'
-                                    as={Link}
-                                    onClick={closeOffcanvas}
-                                >
-                                    <FontAwesomeIcon
-                                        style={DEFAULT_ICON_STYLE}
-                                        icon={faUser}
-                                    />
-                                    Profil
-                                </Nav.Link>
-                                {hasSome(Api.User.RoleEnum.ADMIN) && (
-                                    <Nav.Link
-                                        to='/users'
-                                        as={Link}
-                                        onClick={closeOffcanvas}
-                                    >
-                                        <FontAwesomeIcon
-                                            style={DEFAULT_ICON_STYLE}
-                                            icon={faUsers}
-                                        />
-                                        Používatelia
-                                    </Nav.Link>
-                                )}
-                                <Nav.Link
-                                    to='/categories'
-                                    as={Link}
-                                    onClick={closeOffcanvas}
-                                >
-                                    <FontAwesomeIcon
-                                        style={DEFAULT_ICON_STYLE}
-                                        icon={faList}
-                                    />
-                                    Kategórie
-                                </Nav.Link>
-                                <Nav.Link
-                                    to='/units'
-                                    as={Link}
-                                    onClick={closeOffcanvas}
-                                >
-                                    <FontAwesomeIcon
-                                        style={DEFAULT_ICON_STYLE}
-                                        icon={faScaleBalanced}
-                                    />
-                                    Jednotky
-                                </Nav.Link>
-                                <Nav.Link
-                                    to='/tags'
-                                    as={Link}
-                                    onClick={closeOffcanvas}
-                                >
-                                    <FontAwesomeIcon
-                                        style={DEFAULT_ICON_STYLE}
-                                        icon={faTags}
-                                    />
-                                    Značky
-                                </Nav.Link>
-                                {hasSome(Api.User.RoleEnum.ADMIN, Api.User.RoleEnum.CREATOR) && (
-                                    <Nav.Link
-                                        to='/recipe/create'
-                                        as={Link}
-                                        onClick={closeOffcanvas}
-                                    >
-                                        <FontAwesomeIcon
-                                            style={DEFAULT_ICON_STYLE}
-                                            icon={faPizzaSlice}
-                                        />
-                                        Pridať recept
-                                    </Nav.Link>
-                                )}
-                                <hr />
-                                <Nav.Link
-                                    to={`/recipes?${getDefaultSearchParams()}`}
-                                    as={Link}
-                                    onClick={closeOffcanvas}
-                                >
-                                    <FontAwesomeIcon
-                                        style={DEFAULT_ICON_STYLE}
-                                        icon={faUtensils}
-                                    />
-                                    Všetky recepty
-                                </Nav.Link>
-                                <hr />
-                                {listOfCategories?.map((category) => (
-                                    <Nav.Link
-                                        to={`/recipes?${getDefaultSearchParams(category.id)}`}
-                                        as={Link}
-                                        onClick={closeOffcanvas}
-                                        key={category.id}
-                                    >
-                                        {category.name}
-                                    </Nav.Link>
-                                ))}
-                            </Nav>
-                        </Offcanvas.Body>
-                    </Navbar.Offcanvas>
-                </Container>
-            </Navbar>
-            <Modal
-                show={!!error}
-                message={error}
-                type='error'
-                onClose={() => {
-                    setError(undefined);
-                }}
-            />
-            <Spinner show={isFetchingListOfCategories} />
-        </>
-    );
-};
-
-const AppRoutes: React.FC<{ isLoggedIn: boolean }> = React.memo(({ isLoggedIn }) => {
-    const fallback = isLoggedIn ? '/recipes' : '/signIn';
-
-    const routes = useMemo(
-        () =>
-            isLoggedIn ? (
-                <Route
-                    path='/'
-                    element={
-                        <BookmarkContextProvider>
-                            <Outlet />
-                        </BookmarkContextProvider>
-                    }
-                >
-                    <Route
-                        index={true}
-                        element={
-                            <Navigate
-                                replace
-                                to={'recipes'}
-                            />
-                        }
-                    />
-                    <Route
-                        path='profile'
-                        element={<ProfilePage />}
-                    />
-                    <Route
-                        path='users'
-                        element={<UsersPage />}
-                    />
-                    <Route
-                        path='user'
-                        element={<UserPage />}
-                    />
-                    <Route
-                        path='user/:id'
-                        element={<UserPage />}
-                    />
-                    <Route
-                        path='categories'
-                        element={<CategoriesPage />}
-                    />
-                    <Route
-                        path='category'
-                        element={<CategoryPage />}
-                    />
-                    <Route
-                        path='category/:id'
-                        element={<CategoryPage />}
-                    />
-                    <Route
-                        path='tags'
-                        element={<TagsPage />}
-                    />
-                    <Route
-                        path='tag'
-                        element={<TagPage />}
-                    />
-                    <Route
-                        path='tag/:id'
-                        element={<TagPage />}
-                    />
-                    <Route
-                        path='units'
-                        element={<UnitCategoriesPage />}
-                    />
-                    <Route
-                        path='unit/:categoryId/'
-                        element={<UnitPage />}
-                    />
-                    <Route
-                        path='unit/:categoryId/:unitId'
-                        element={<UnitPage />}
-                    />
-                    <Route
-                        path='unitCategory'
-                        element={<UnitCategoryPage />}
-                    />
-                    <Route
-                        path='unitCategory/:id'
-                        element={<UnitCategoryPage />}
-                    />
-                    <Route
-                        path='recipe/create'
-                        element={<RecipeEditPage />}
-                    />
-                    <Route
-                        path='recipe/edit/:recipeId'
-                        element={<RecipeEditPage />}
-                    />
-                    <Route
-                        path='recipe/:recipeId'
-                        element={<RecipeViewPage />}
-                    />
-                    <Route
-                        path='recipes'
-                        element={<RecipeSearchPage />}
-                    />
-                </Route>
-            ) : (
-                <Route
-                    path='/signIn'
-                    element={<SignInPage />}
-                />
-            ),
-        [isLoggedIn]
-    );
-
-    return (
-        <Container
-            as={'main'}
-            className='py-4'
-        >
-            <Routes>
-                <Route
-                    path='/confirm'
-                    element={<ConfirmationPage />}
-                />
-                <Route
-                    path='/confirm/:username/:key'
-                    element={<ConfirmationPage />}
-                />
-                <Route
-                    path='/resetRequest'
-                    element={<ResetPasswordRequestPage />}
-                />
-                <Route
-                    path='/reset'
-                    element={<ResetPasswordPage />}
-                />
-                <Route
-                    path='/reset/:username/:key'
-                    element={<ResetPasswordPage />}
-                />
-                <Route
-                    path='/consent'
-                    element={<ConsentPage />}
-                />
-                {routes}
-                <Route
-                    path='*'
-                    element={
-                        <Navigate
-                            replace
-                            to={fallback}
-                        />
-                    }
-                />
-            </Routes>
-        </Container>
-    );
-});
+import useRouter from './hooks/use-router';
 
 const App: React.FC = () => {
-    const { isLoggedIn, setHasCookieConsent } = useContext(AuthContext);
+    const { isLoggedIn } = useContext(AuthContext);
+    const router = useRouter(isLoggedIn);
 
-    const [error, setError] = useState<string>();
-
-    const {
-        data: userInfo,
-        isFetching: isFetchingUserInfo,
-        error: userInfoError
-    } = useQuery({
-        queryKey: ['currentUser'],
-        queryFn: ({ signal }) => authApi.user({ signal }),
-        enabled: isLoggedIn
-    });
-
-    useEffect(() => {
-        if (userInfoError) {
-            formatErrorMessage(userInfoError).then((message) => setError(message));
-        }
-    }, [userInfoError]);
-
-    let username = `${userInfo?.firstName ?? ''} ${userInfo?.lastName ?? ''}`.trim();
-    if (!username) {
-        username = userInfo?.username ?? '';
-    }
-
-    const onAcceptHandler = useCallback(() => {
-        setHasCookieConsent(true);
-    }, [setHasCookieConsent]);
-
-    const onDeclineHandler = useCallback(() => {
-        setHasCookieConsent(false);
-    }, [setHasCookieConsent]);
-
-    return (
-        <>
-            <CookieConsent
-                location='bottom'
-                buttonText='Súhlasím'
-                declineButtonText='Nesúhlasím'
-                containerClasses='bg-primary'
-                buttonClasses='btn btn-warning rounded m-2'
-                declineButtonClasses='btn btn-danger rounded m-2'
-                buttonWrapperClasses='d-flex flex-column'
-                sameSite='Lax'
-                cookieSecurity={false}
-                enableDeclineButton={true}
-                disableButtonStyles={true}
-                setDeclineCookie={false}
-                flipButtons={true}
-                overlay={true}
-                onAccept={onAcceptHandler}
-                onDecline={onDeclineHandler}
-            >
-                <h2>Cookies</h2>
-                <p>
-                    Udelením súhlasu umožníte aplikácií ukladať informácie do lokálneho úložiska (localStorage) prehliadača. Aplikácia po
-                    udelení súhlasu umožňuje zapamätať si práve prihláseného používateľa. Vďaka tomu sa nebudete musieť pri opätovnej
-                    návšteve prihlasovať. Lokálne úložisko sa zároveň môže používať na uloženie "záložiek" na recepty. Uložené záložky sa
-                    viažú na prehliadač a nie používateľské konto. Po odhlásení budú tieto údaje odstránené.
-                </p>
-                <p>Informácia o udelenom súhlase sa ukladá do cookies prehliadača a je platná na 1 rok od udelenia súhlasu.</p>
-                <p>
-                    Svoje rozhodnutie môžete zmeniť v{' '}
-                    <Link
-                        to={'/consent'}
-                        className='text-white text-decoration-underline'
-                    >
-                        nastaveniach cookies
-                    </Link>{' '}
-                    alebo po prihlásení v profile používateľa.
-                </p>
-            </CookieConsent>
-            <Navigation
-                isLoggedIn={isLoggedIn}
-                username={username}
-            />
-            <AppRoutes isLoggedIn={isLoggedIn} />
-            <Modal
-                show={!!error}
-                message={error}
-                type='error'
-                onClose={() => {
-                    setError(undefined);
-                }}
-            />
-            <Spinner show={isFetchingUserInfo} />
-        </>
-    );
+    return <RouterProvider router={router} />;
 };
 
 export default App;
