+++ src/components/recipes/RecipeThumbnail.tsx
