+++ src/components/recipes/edit/Ingredients/Ingredient.tsx
@@ -4,7 +4,7 @@ import { get } from 'lodash';
 import React, { useCallback, useId, useState } from 'react';
 import { Button, Form, InputGroup } from 'react-bootstrap';
 import { UseFieldArrayMove, UseFieldArrayRemove, useFormContext } from 'react-hook-form';
-import { SelectGroupOptions } from '../../../UI/Select';
+import { SelectGroupOptions } from '../../../ui/Select';
 
 type IngredientProps = {
     name: string;
