+++ src/components/recipes/edit/Ingredients/Ingredients.tsx
@@ -3,7 +3,7 @@ import { Button, Form, Stack } from 'react-bootstrap';
 import { useFieldArray, useFormContext } from 'react-hook-form';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCirclePlus } from '@fortawesome/free-solid-svg-icons';
-import { SelectGroupOptions } from '../../../UI/Select';
+import { SelectGroupOptions } from '../../../ui/Select';
 import Ingredient from './Ingredient';
 import { RecipeForm } from '../../../../pages/recipes/RecipeEditPage';
