+++ src/components/recipes/view/PictureView.tsx
