+++ src/index.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { BrowserRouter } from 'react-router-dom';
 import './index.scss';
-import App from './App';
 import AuthContextProvider from './store/auth-context';
+import App from './App';
 
 const queryClient = new QueryClient({
     defaultOptions: {
@@ -20,11 +19,12 @@ root.render(
     <React.StrictMode>
         <QueryClientProvider client={queryClient}>
             <AuthContextProvider>
-                <BrowserRouter>
-                    <App />
-                </BrowserRouter>
+                <App />
             </AuthContextProvider>
-            <ReactQueryDevtools initialIsOpen={false} />
+            <ReactQueryDevtools
+                initialIsOpen={false}
+                buttonPosition='bottom-left'
+            />
         </QueryClientProvider>
     </React.StrictMode>
 );
