+++ src/pages/signIn/ResetPasswordRequestPage.tsx
@@ -6,9 +6,9 @@ import * as yup from 'yup';
 import { Api } from '../../openapi';
 import { formatErrorMessage } from '../../utils/errorMessages';
 import { authApi } from '../../utils/apiWrapper';
-import Input from '../../components/UI/Input';
-import Modal from '../../components/UI/Modal';
-import Spinner from '../../components/UI/Spinner';
+import Input from '../../components/ui/Input';
+import Modal from '../../components/ui/Modal';
+import Spinner from '../../components/ui/Spinner';
 
 const schema = yup
     .object({
